requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
attrs==19.3.0
autopep8==1.5.3
Babel==2.8.0
backcall==0.2.0
beautifulsoup4==4.9.1
CacheControl==0.12.6
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0.post1
cryptography==2.9.2
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
distro==1.5.0
dj-database-url==0.5.0
Django==3.0.7
django-allauth==0.42.0
django-bootstrap4==2.1.1
django-cors-headers==3.4.0
django-crispy-forms==1.9.1
django-debug-toolbar==2.2
django-environ==0.4.5
django-extensions==3.0.1
django-filter==2.3.0
django-haystack==2.8.1
django-js-asset==1.2.2
django-mptt==0.11.0
django-phonenumber-field==4.0.0
django-seed==0.2.2
django-session-security==2.6.6
django-taggit==1.3.0
django-tinymce4-lite==1.8.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
drf-nested-routers==0.91
elasticsearch==7.8.0
elasticsearch-dsl==7.2.1
et-xmlfile==1.0.1
Faker==4.1.1
flake8==3.8.3
gunicorn==20.0.4
html5lib==1.1
httpie==2.2.0
idna==2.10
importlib-metadata==1.7.0
ipaddr==2.2.0
ipdb==0.13.3
ipython==7.16.1
ipython-genutils==0.2.0
isort==4.3.21
jdcal==1.4.1
jedi==0.17.1
jsmin==2.2.2
lazy-object-proxy==1.5.0
lockfile==0.12.2
mccabe==0.6.1
more-itertools==8.4.0
msgpack==1.0.0
numpy==1.19.0
oauthlib==3.1.0
openpyxl==3.0.4
packaging==20.4
parso==0.7.0
pep517==0.8.2
pexpect==4.8.0
phonenumbers==8.12.6
pickleshare==0.7.5
Pillow==7.2.0
pip-review==1.1.0
pluggy==0.13.1
progress==1.5
prompt-toolkit==3.0.5
psycopg2==2.8.5
ptyprocess==0.6.0
pur==5.3.0
py==1.9.0
pycodestyle==2.6.0
pycparser==2.20
pydot==1.4.1
pyflakes==2.2.0
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.5.3
pyparsing==2.4.7
pytest==5.4.3
pytest-django==3.9.0
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.2.0
pytoml==0.1.21
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.1
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.25.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.1.0
wrapt==1.12.1
zipp==3.1.0